Revert "Merge origin/master"
This reverts commit 605ad32e8ed67228064200a2b11f4890bc848ce6.

+++ src/helpers/Clock.java
@@ -53,7 +53,7 @@ public class Clock {
         TotalTime += d;
     }
     
-    public static void ChangeMultiplier(int Change)
+    public static void ChangeMultiplier(float Change)
     {
         if(Multiplier + Change < -1 && Multiplier + Change > 7)
         {
